fix(ecology): type key schemas and plot-effect configs
- preserve typed array statics in TypedArraySchemas
- align feature/plot-effect key schemas with Static typing
- type plot-effects configs and fix rule imports

+++ mods/mod-swooper-maps/src/domain/ecology/ops/plan-feature-placements/rules/adjacency.ts
@@ -1,5 +1,5 @@
 export function hasAdjacentFeatureType(
-  featureField: Int32Array,
+  featureField: Int16Array,
   width: number,
   height: number,
   x: number,
